+++ modules/circe/src/test/scala/io/bartholomews/fsclient/samples/v2/ImplicitGrantExample.scala
@@ -11,7 +11,7 @@ object ImplicitGrantExample extends App {
 
   def dealWithIt = throw new Exception("¯x--(ツ)--x")
 
-  implicit val backend: SttpBackend[Identity, Any] = HttpURLConnectionBackend()
+  val backend: SttpBackend[Identity, Any] = HttpURLConnectionBackend()
 
   val userAgent: UserAgent = UserAgent(
     appName = "SAMPLE_APP_NAME",
@@ -27,7 +27,7 @@ object ImplicitGrantExample extends App {
 
   val myRedirectUri = RedirectUri(uri"https://my-app/callback")
 
-  val client = FsClient.v2.clientPassword(userAgent, ClientPasswordAuthentication(myClientPassword))
+  val client = FsClient.v2.clientPassword(userAgent, ClientPasswordAuthentication(myClientPassword))(backend)
 
   // 1. Prepare an authorization token request
   val authorizationTokenRequest = AuthorizationTokenRequest(