+++ modules/core/src/main/scala/io/bartholomews/fsclient/core/FsApiClient.scala
@@ -5,5 +5,5 @@ import io.bartholomews.fsclient.core.oauth.Signer
 import sttp.client3.SttpBackend
 
 abstract case class FsApiClient[F[_], S <: Signer](client: FsClient[F, S]) extends FsClientSttpExtensions {
-  implicit val backend: SttpBackend[F, Any] = client.backend
+  val backend: SttpBackend[F, Any] = client.backend
 }