+++ modules/core/src/test/scala/io.bartholomews.fsclient/client/IdentityClient.scala
@@ -7,11 +7,11 @@ import sttp.client3.{HttpURLConnectionBackend, Identity, SttpBackend}
 
 trait IdentityClient {
 
-  implicit val sttpIdentityBackend: SttpBackend[Identity, Any] = HttpURLConnectionBackend()
+  val backend: SttpBackend[Identity, Any] = HttpURLConnectionBackend()
 
   val clientNoAuth: FsClient[Identity, AuthDisabled.type] =
-    FsClient(sampleUserAgent, AuthDisabled, sttpIdentityBackend)
+    FsClient(sampleUserAgent, AuthDisabled, backend)
 
   val clientCredentials: FsClient[Identity, SignerV1] =
-    FsClient.v1.clientCredentials(sampleUserAgent, sampleConsumer)
+    FsClient.v1.clientCredentials(sampleUserAgent, sampleConsumer)(backend)
 }