+++ modules/core/src/test/scala/io.bartholomews.fsclient/core/oauth/v1/OAuthV1Test.scala
@@ -34,7 +34,7 @@ class OAuthV1Test extends AnyWordSpec with IdentityClient with WiremockServer wi
         serverUri = uri"$wiremockBaseUri/oauth/request-token",
         sampleUserAgent,
         ResourceOwnerAuthorizationUri(uri"https://some-server/oauth/authorize")
-      )
+      )(backend)
 
     "get temporary credentials" in {